+++ tests/Feature/ModularInstallCommandTest.php
@@ -31,7 +31,8 @@ it('can install modular and configure vite automatically', function () {
 
     $this->artisan('modular:install')
         ->expectsConfirmation('Would you like to publish the modular stubs for customization?', 'yes')
-        ->expectsConfirmation('Would you like to automatically configure composer.json for modular autoloading?', 'yes')
+        ->expectsConfirmation("Would you like to add optimized PSR-4 autoloading for 'Modules\\' to your composer.json?", 'yes')
+        ->expectsConfirmation('Would you like to automatically configure it?', 'yes')
         ->expectsConfirmation('Would you like to automatically configure vite.config.js?', 'yes')
         ->expectsConfirmation('Would you like to show some love by starring the repo on GitHub? ⭐', 'no')
         ->assertExitCode(0);
@@ -62,7 +63,8 @@ it('can install modular and show manual configuration if declined', function () 
 
     $this->artisan('modular:install')
         ->expectsConfirmation('Would you like to publish the modular stubs for customization?', 'no')
-        ->expectsConfirmation('Would you like to automatically configure composer.json for modular autoloading?', 'no')
+        ->expectsConfirmation("Would you like to add optimized PSR-4 autoloading for 'Modules\\' to your composer.json?", 'no')
+        ->expectsConfirmation('Would you like to automatically configure it?', 'no')
         ->expectsConfirmation('Would you like to automatically configure vite.config.js?', 'no')
         ->expectsConfirmation('Would you like to show some love by starring the repo on GitHub? ⭐', 'no')
         ->assertExitCode(0);